@@ -1,11 +1,9 @@
 package com.kararnab.contacts.widgets;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.AttributeSet;
 import android.view.View;
-
-import com.kararnab.contacts.UiUtils;
 
 import org.jetbrains.annotations.Nullable;
 
